+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 This is used for recording Invader's changes. This changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Untagged
+### Fixed
+- invader-bludgeon: Fixed missing-bitmap-sequences applying to sprites
+- invader-build: Fixed some empty sequences being put in maps for sprites
+
 ## [0.50.5] - 2022-08-14
 ### Added
 - Added continuous_damage_effect.